Added token provider and range

+++ Api/src/UrlShortener.Core/Base62EncodingExtensions.cs
@@ -6,14 +6,15 @@ public static class Base62EncodingExtensions
         "0123456789" +
         "ABCDEFGHIJKLMNOPQRSTUVWXYZ" +
         "abcdefghijklmnopqrstuvwxyz";
-    public static string EncodeToBase62(this int number)
+    
+    public static string EncodeToBase62(this long number)
     {
         if (number == 0) return Alphanumeric[0].ToString();
 
         var result = new Stack<char>();
         while (number > 0)
         {
-            result.Push(Alphanumeric[number % 62]);
+            result.Push(Alphanumeric[(int)number % 62]);
             number /= 62;
         }
         return new string(result.ToArray());
